Created initial basic hashmap and performed benchmarking
Initial benchmarking showed considerably worse performance in custom hashmap. On this particular version, the difference was approx ~330% slower.
Removal of the setup from the benchmark reduced the time by ~40%. The same operation with the built-in hashmap had ~20% redduction in time. This needs to be improved.
Casting between types had a major impact.
Changing []byte types on keys to strings and changing the hash function to hash the strings directly resulted in ~25% reduction in time.
Changing interface{} to string for values resulted in a further ~25% reduction in time.